FEXRootFSFetcher: ExecAndWaitForResponse sign extend program result
Only the lower 8bits of the execve result is the program result.
Makes sure to sign extend it so -1 is a true -1 instead of 255

+++ Source/Tools/FEXRootFSFetcher/Main.cpp
@@ -20,7 +20,7 @@ namespace Exec {
       int32_t Status{};
       waitpid(pid, &Status, 0);
       if (WIFEXITED(Status)) {
-        return WEXITSTATUS(Status);
+        return (int8_t)WEXITSTATUS(Status);
       }
     }
